requirements.txt
# Web Framework
Flask==2.3.3
Jinja2==3.1.2

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2

# Google Trends API
pytrends==4.9.2

# Data Analysis
pandas==2.1.3
numpy==1.26.2

# NLP
nltk==3.8.1
spacy==3.7.2

# Visualization
matplotlib==3.8.2
plotly==5.18.0

# Testing
pytest==7.4.3

# Utilities
python-dotenv==1.0.0

